requirements.txt
selenium==4.17.2
webdriver-manager==4.0.1
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
python-dateutil==2.8.2
